requirements.txt
pyTelegramBotAPI
Pyrogram==1.4.16
tgcrypto
asyncio
DateTime
aiofiles
motor
pymongo
umongo
dnspython
psutil
python-arq
requests
aiohttp
telegraph
opencv-python-headless
Pillow
speedtest-cli
